+++ src/Exception/VoipNowException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace HyderKamran\VoipNow\Exception;
+
+use RuntimeException;
+
+class VoipNowException extends RuntimeException
+{
+}