+++ src/Interface/ConnectorInterface.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace HyderKamran\VoipNow\Interface;
-interface ConnectorInterface
-{
-    public function connect(array $config);
-}